+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 Changes:
 - cshaper: add simple tc-cake based shaping
 
+Fixes:
+- pyroute2: workaround NetlinkError regression (pyroute2 #845 #847)
+
 ## 1.5.6 - 2021-09-25
 
 Changes: